Test without blocking http server
